Move construct_relative_lambda_schedule to attic
Unused, and is not symmetric (`schedule != (1 - schedule[::-1])`)
